Added multiplayer methods
Added new methods that will support the implementation of a multiplayer mode. These methods are modified versions of the single player methods. These new methods should eventually replace the single player methods. I did not remove any of the old methods, so this should not affect any of the single player tests.